refactor: update handleKeyUp method
